+++ inc/Functions.hpp
@@ -28,6 +28,6 @@ void fill_string_spaces(string & str);
 void lowercase(string& str);
 void clear_fail_state();
 vector<string> split(const string &s, char delim);
-vector<pair<string, int>> order(map<string, int> records);
+vector<pair<string, float>> order(map<string, float> records);
 
 #endif
