+++ src/Functions.cpp
@@ -35,10 +35,10 @@ vector<string> split (const string &s, char delim) {
     return result;
 }
 
-vector<pair<string, int>> order (map<string, int> records){
-    vector<pair<string, int>> sorted_vector;
-    copy(records.begin(), records.end(), back_inserter<vector<pair<string, int>>>(sorted_vector));
-    sort(sorted_vector.begin(), sorted_vector.end(),[](const pair<string, int>& l, const pair<string, int>& r) {
+vector<pair<string, float>> order (map<string, float> records){
+    vector<pair<string, float>> sorted_vector;
+    copy(records.begin(), records.end(), back_inserter<vector<pair<string, float>>>(sorted_vector));
+    sort(sorted_vector.begin(), sorted_vector.end(),[](const pair<string, float>& l, const pair<string, float>& r) {
         if (l.second != r.second)
             return l.second > r.second;
         return l.first > r.first;
